+++ HISTORY.rst
@@ -8,6 +8,8 @@ Changelog
 - Add "speed" config option.
 - Fix short option for "--shell".
 - Custom prompts are colored.
+- Remove unnecessary --check-output option, which was only used for testing.
+- Fix bug where cwd would not update in custom prompts.
 
 0.1.0 (2014-06-16)
 ------------------